+++ app/components/Admin.jsx
@@ -50,6 +50,7 @@ export class Admin extends React.Component {
                         <Link to={'/admin/pages'} className="expanded button"><i className="fi-page-multiple"></i> Pages</Link>
                         <Link to={'/admin/players'} className="expanded button"><i className="fi-torsos-all"></i> Players</Link>
                         <Link to={'/admin/banner'} className="expanded button"><i className="fi-layout"></i> Banner</Link>
+                        <Link to={'/admin/sponsors'} className="expanded button"><i className="fi-pricetag-multiple"></i> Sponsors</Link>
                     </ul>
                 </div>
             );
